+++ index.d.ts
@@ -1,3 +1,4 @@
 import Switch from './components/switch';
 import Header from './components/header';
-export { Switch, Header, };
+import StarRating from './components/starrating';
+export { Switch, Header, StarRating, };