+++ index.js
@@ -4,3 +4,5 @@ const switch_1 = require("./components/switch");
 exports.Switch = switch_1.default;
 const header_1 = require("./components/header");
 exports.Header = header_1.default;
+const starrating_1 = require("./components/starrating");
+exports.StarRating = starrating_1.default;